refactor: go back to normal nix version
The main issue I had with it is that it overrides `netrc` file for use
with its CLI and FlakeHub, and doesn't allow any way to add custom
entries to it, and since I have a private binary cache which requires
auth using the `netrc` file it caused problems. Soo... long story short.
I've decided to remove it.